requirements.txt
pybit>=5.8.0
httpx>=0.27.0
python-dotenv>=1.0.1
